Enhancement: Extract AutoReview test-suite and split tests
