+++ .travis.yml
@@ -62,6 +62,7 @@ jobs:
         - mysqldump -uroot $TRAVIS_DB > tests/dump.sql
 
       script:
+        - vendor/bin/phpunit --testsuite auto-review
         - vendor/bin/phpunit --testsuite integration
         - if [[ "$WITH_COVERAGE" == "true" ]]; then xdebug-enable; fi
         - if [[ "$WITH_COVERAGE" == "true" ]]; then vendor/bin/phpunit --testsuite unit --coverage-clover build/logs/clover.xml; else vendor/bin/phpunit --testsuite unit; fi
