+++ phpunit.xml.dist
@@ -27,6 +27,9 @@
         </whitelist>
     </filter>
     <testsuites>
+        <testsuite name="auto-review">
+            <directory>tests/AutoReview</directory>
+        </testsuite>
         <testsuite name="unit">
             <directory>tests/Unit</directory>
         </testsuite>
